chore: various refactor following code review
